[WIP] use NavController for navigation and refactor to remove fragments
