+++ app/build.gradle
@@ -44,6 +44,7 @@ dependencies {
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.5.1'
     implementation 'androidx.lifecycle:lifecycle-livedata-ktx:2.5.1'
+    implementation "androidx.navigation:navigation-compose:2.5.1"
     implementation 'com.google.android.material:material:1.7.0'
 
     implementation 'androidx.compose.ui:ui:1.1.1'
