+++ app/src/main/java/com/github/mrpaulblack/personalrecipes/MainActivity.kt
@@ -2,16 +2,35 @@ package com.github.mrpaulblack.personalrecipes
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import androidx.activity.compose.setContent
+import androidx.activity.viewModels
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.material.Button
+import androidx.compose.material.Scaffold
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.livedata.observeAsState
+import androidx.compose.ui.Alignment
+import androidx.navigation.NavController
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import androidx.navigation.compose.rememberNavController
+import com.github.mrpaulblack.personalrecipes.ui.counter.CounterView
+import com.github.mrpaulblack.personalrecipes.ui.main.MainView
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-
-        supportFragmentManager.beginTransaction().replace(R.id.container, CounterFragment()).commit();
-    }
-
-    override fun onBackPressed() {
-        supportFragmentManager.popBackStack();
+        setContent {
+            val navController = rememberNavController()
+            Scaffold() {
+                NavHost(navController, startDestination = "main") {
+                    composable("main") { MainView.Content(navController) }
+                    composable("counter") { CounterView.Content() }
+                }
+            }
+        }
     }
 }
