+++ app/src/main/java/com/github/mrpaulblack/personalrecipes/ui/counter/CounterViewModel.kt
@@ -1,12 +1,12 @@
-package com.github.mrpaulblack.personalrecipes
+package com.github.mrpaulblack.personalrecipes.ui.counter
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
 class CounterViewModel: ViewModel() {
-    val counter = MutableLiveData(0);
+    val counter = MutableLiveData(0)
 
     fun onCounterClicked() {
-        counter.value = (counter.value ?: 0) +1;
+        counter.value = (counter.value ?: 0) +1
     }
 }
